Create Section 4, Lecture 35 Form Exam
In this file, I created a simple form where you register for a website. You have to put in your name, your email, your gender, a password, and a date of birth. Then you need to accept the terms and submit. Placeholders are in the file, some data points are required, and the password must be between 5 and 10 characters.